Start a new request to Brain backend for each RECOGNIZE request.
This is necessary mainly because Brain does not (at least not
currently) support client-side endpointing, so we cannot line up the
ASR results with the lifetime of the RECOGNIZE request without doing
some timestamp bookkeeping (which has its own edge cases and
ambiguities). Therefore, we just start a new request.

Starting a request should be fast enough, and if we want it to be a
little bit faster, we can make this plugin acquire and refresh an auth
token, which is a bit more efficient than authenticating using passwords.