Eliminate the UniMRCP task thread in favour of the Tokio scheduler.
This accomplishes a number of things.

It reduces the total number of system threads running. It also reduces
the number of independently spawned tokio tasks, since the entire
WebSocket request call can be run in a single async function.

It eliminates the messaging system introduced by the UniMRCP task
thread. Although I can make some guesses based on the UniMRCP source
code and tracing in a debugger, the memory management of messages is
not clearly documented, so it's probably better to not rely on
it. Additionally, this means that running tasks on a worker thread is
no longer an `unsafe` operation.

It also eliminates the back pointer from channel back to the engine,
since the channel no longer needs to access the engine's `task`
pointer. Instead, channels hold on to a tokio runtime handle.