Add support for using DG's VAD instead of UniMRCP's VAD.
There is a tradeoff here. On one hand, using DG's VAD will require
that an actual _word_ is recognized, which means that START-OF-SPEECH
will not be sent just because of some non-speech noise. However, this
also incurs the additional latency of waiting for those speech
results, whereas the UniMRCP VAD is much more responsive.