Rewrite Dockerfile to work across additional platforms.
Notable changes include:

OpenSSL is built and statically linked. This should avoid issuses
where some OS's have libssl.so version 1.0, and some have 1.1. I
considered and attempted to build using musl, but it appears that
Rust's musl target doesn't fully support building shared objects
yet (the `cdylib` crate type).

The base image is CentOS 7 instead of the Rust base image. This is
because the Rust images are based on a more recent Debian, and we
would like to build with an older version of glibc.

UniMRCP itself is downloaded as a pre-built package from its package
repository. This avoids the need to build it and its
dependencies. Note that the UniMRCP repo requires authentication,
which is provided as build args to Docker via the `UNIMRCP_USERNAME`
and `UNIMRCP_PASSWORD` variables, which are populated by GitHub
secrets.