Add support for numerals and ner parameters.
This is a quick fix to unblock a user, because we didn't have the
ability to add arbitrary query parameters to the websocket request.

It should probably be followed up with a more general solution, and
the configuration format _might_ change.

For now, numerals and ner can be set in the plugin config as follows:

```xml
<plugin-factory>
  <engine id="Deepgram" name="libdgmrcp" enable="true">
    <param name="brain_url" value="wss://brain.deepgram.com/v2/"/>
    <param name="brain_username" value="{{ USERNAME }}"/>
    <param name="brain_password" value="{{ PASSWORD }}"/>
    <param name="ner" value="true"/>
    <param name="numerals" value="true"/>
  </engine>
</plugin-factory>
```