Version 0.3.6
The main feature of this release is that we immediately and
optimistically respond to a `RECOGNIZE` request with an `IN-PROGRESS`
message. This encourages the client to start sending audio as soon as
possible, which helps to avoid missing the start of speech or a short
utterance while a WebSocket connection is being established. Prior to
this, if an error happened during the connection then the response to
`RECOGNIZE` would contain an error. Now, the error reporting is moved
to the `RECOGNITION-COMPLETE` message. Clients should already be
handling the possibility of failure in this message, so this is not
considered a breaking API change.